+++ DX11Starter/DXCore.cpp
@@ -400,10 +400,10 @@ void DXCore::UpdateTimer()
 	// Calculate delta time and clamp to zero
 	//  - Could go negative if CPU goes into power save mode 
 	//    or the process itself gets moved to another core
-	deltaTime = max((float)((currentTime - previousTime) * perfCounterSeconds), 0.0f);
+	deltaTime = max((double)((currentTime - previousTime) * perfCounterSeconds), 0.0f);
 
 	// Calculate the total time from start to now
-	totalTime = (float)((currentTime - startTime) * perfCounterSeconds);
+	totalTime = (double)((currentTime - startTime) * perfCounterSeconds);
 
 	// Save current time for next frame
 	previousTime = currentTime;
